requirements.txt
python-telegram-bot~=13.7
sqlalchemy==1.4.22
mariadb==1.0.7
pytest==6.2.4
